Create UPDATE SQL statement

Creates an UPDATE SQL statement for a single table. Currently, this function cannot handle using expressions in the set parameter. UPDATE is only useful at this time for single table updates when set contains actual values.

UPDATE(table, set, ..., .ignore = FALSE, .order = NULL, .limit = NULL)

Arguments

table
Table name for update (can be named)

set
Named vector of column-value pairs, where the vector name is the column name.

...
Conditions passed on to WHERE clause (optional)

.ignore
Add IGNORE keyword to UPDATE clause

.order
Optional vector of columns passed to ORDER_BY

.limit
Optional number of rows for LIMIT condition

Details

Improvements are planned for this function to handle multiple table updates and the use of SQL expressions in the SET statement.

Examples

UPDATE('iris', c(some_column = 1, some_other_col = "high"), eq(another_col = 2), geq(a_third_col = 10))

#> [1] "UPDATE iris  SET some_column=1, some_other_col=\"high\" WHERE another_col=2 AND a_third_col>=10"

UPDATE('t1', c(col1 = 'a'))

#> [1] "UPDATE t1  SET col1=\"a\""

UPDATE('t1', c(col1 = 'a', col2 = 42), 'id' %IN% 1:5)

#> [1] "UPDATE t1  SET col1=\"a\", col2=42 WHERE id IN (1, 2, 3, 4, 5)"

UPDATE('t', c(id = 'id + 1'), .order = DESC('id'))

#> [1] "UPDATE t  SET id=\"id + 1\"  ORDER BY id DESC"